PRA-59 save results
