+++ .gitignore
@@ -2,6 +2,7 @@ node_modules/
 .env
 .idea
 .vscode
+*.log
 dist
 ssr-dist
 tmp
