+++ packages/client/src/App.tsx
@@ -6,7 +6,7 @@ import styles from './styles.module.scss';
 function App() {
   useEffect(() => {
     const fetchServerData = async () => {
-      const url = `http://localhost:${__SERVER_PORT__}`;
+      const url = `http://localhost:${__SERVER_PORT__ || 3001}/api`;
       const response = await fetch(url);
       const data = await response.json();
       console.log(data);
