--- packages/client/src/pages/ErrorPage/index.ts
@@ -1,5 +1,0 @@
-import { lazy } from 'react';
-
-export const ErrorPage = lazy(
-  () => import('./ErrorPage')
-);
