--- packages/client/src/pages/ForumChatPage/index.ts
@@ -1,3 +1,0 @@
-import { lazy } from 'react';
-
-export const ForumChatPage = lazy(() => import('./ForumChatPage'));
