--- packages/client/src/pages/ForumPage/index.ts
@@ -1,3 +1,0 @@
-import { lazy } from 'react';
-
-export const ForumPage = lazy(() => import('./ForumPage'));
