--- packages/client/src/pages/GamePage/GamePage.tsx
@@ -1,6 +1,0 @@
-import { FunctionComponent } from 'react';
-import { GameBlock } from './Components/GameBlock';
-
-const GamePage = () => <GameBlock />;
-
-export default GamePage as FunctionComponent;
