--- packages/client/src/pages/GamePage/index.ts
@@ -1,3 +1,0 @@
-import { lazy } from 'react';
-
-export const GamePage = lazy(() => import('./GamePage'));
