+++ packages/client/src/pages/GamePage/index.tsx
@@ -1,0 +1,3 @@
+import { GameBlock } from './Components/GameBlock';
+
+export const GamePage = () => <GameBlock />;
