--- packages/client/src/pages/Layout/index.ts
@@ -1,1 +1,0 @@
-export * from './Layout';
