--- packages/client/src/pages/LeaderboardPage/index.ts
@@ -1,3 +1,0 @@
-import { lazy } from 'react';
-
-export const LeaderboardPage = lazy(() => import('./LeaderboardPage'));
