--- packages/client/src/pages/MainPage/index.ts
@@ -1,3 +1,0 @@
-import { lazy } from 'react';
-
-export const MainPage = lazy(() => import('./MainPage'));
