--- packages/client/src/pages/ProfilePage/index.ts
@@ -1,3 +1,0 @@
-import { lazy } from 'react';
-
-export const ProfilePage = lazy(() => import('./ProfilePage'));
