--- packages/client/src/pages/SigninPage/index.ts
@@ -1,3 +1,0 @@
-import { lazy } from 'react';
-
-export const SigninPage = lazy(() => import('./SigninPage'));
