--- packages/client/src/pages/SignupPage/index.ts
@@ -1,3 +1,0 @@
-import { lazy } from 'react';
-
-export const SignupPage = lazy(() => import('./SignupPage'));
