--- packages/client/src/pages/StartPage/index.ts
@@ -1,3 +1,0 @@
-import { lazy } from 'react';
-
-export const StartPage = lazy(() => import('./StartPage'));
