+++ packages/client/src/pages/StartPage/startPage.module.scss
@@ -1,5 +1,3 @@
-@import 'style_constants';
-
 .header {
   display: flex;
   width: 90%;
