--- packages/client/src/pages/UIPage/index.ts
@@ -1,3 +1,0 @@
-import { lazy } from 'react';
-
-export const UIPage = lazy(() => import('./UIPage'));
