+++ packages/server/package.json
@@ -27,6 +27,7 @@
     "@typescript-eslint/eslint-plugin": "^5.35.1",
     "@typescript-eslint/parser": "^5.35.1",
     "babel-jest": "^29.0.1",
+    "browser-env": "^3.3.0",
     "eslint": "^8.23.0",
     "jest": "^28",
     "nodemon": "^2.0.19",
