+++ packages/server/tsconfig.prod.json
@@ -1,4 +1,4 @@
 {
   "extends": "./tsconfig",
-  "exclude": ["**/*.test.ts", "**/*.mock.ts", "**/__test__"]
+  "exclude": ["**/*.test.ts", "**/*.mock.ts", "**/__test__", "**/*.d.ts"]
 }
